Add files via upload
Welcome user to this software, in this program you can enter and retrieve the data of students. First you will be asked to enter number of students and then the Roll Number, Name, Department, Course, year Of Joining  respectively. Then you will be asked to enter the year to get the list of names of student who joined in that particular year. Then at the end you will be asked to enter the roll number of student to fetch the data of that student.